+++ src/raygun.rum.js
@@ -7,7 +7,7 @@
  */
 
 var raygunRumFactory = function (window, $, Raygun) {
-    Raygun.RealUserMonitoring = function (apiKey, apiUrl, makePostCorsRequest, user, version, excludedHostNames, excludedUserAgents, debugMode) {
+    Raygun.RealUserMonitoring = function (apiKey, apiUrl, makePostCorsRequest, user, version, excludedHostNames, excludedUserAgents, debugMode, maxVirtualPageDuration) {
         var self = this;
         var _private = {};
 
@@ -17,6 +17,7 @@ var raygunRumFactory = function (window, $, Raygun) {
         this.debugMode = debugMode;
         this.excludedHostNames = excludedHostNames;
         this.excludedUserAgents = excludedUserAgents;
+        this.maxVirtualPageDuration = maxVirtualPageDuration || 1800000; // 30 minutes
 
         this.makePostCorsRequest = function (url, data) {
             if (self.excludedUserAgents instanceof Array) {
@@ -48,6 +49,7 @@ var raygunRumFactory = function (window, $, Raygun) {
             makePostCorsRequest(url, data);
         };
         this.sessionId = null;
+        this.virtualPage = null;
         this.user = user;
         this.version = version;
         this.heartBeatInterval = null;
@@ -58,7 +60,11 @@ var raygunRumFactory = function (window, $, Raygun) {
                 self.pageLoaded(isNewSession);
             });
 
-            window.onbeforeunload = function () {
+            var clickHandler = function () {
+                this.updateCookieTimestamp();
+            }.bind(_private);
+            
+            var unloadHandler = function () {
                 var data = [];
 
                 extractChildData(data);
@@ -80,10 +86,6 @@ var raygunRumFactory = function (window, $, Raygun) {
                 }
             };
 
-            var clickHandler = function () {
-                this.updateCookieTimestamp();
-            }.bind(_private);
-
             var visibilityChangeHandler = function () {
                 if (document.visibilityState === 'visible') {
                     this.updateCookieTimestamp();
@@ -94,6 +96,7 @@ var raygunRumFactory = function (window, $, Raygun) {
             if (window.addEventListener) {
                 window.addEventListener('click', clickHandler);
                 document.addEventListener('visibilitychange', visibilityChangeHandler);
+                window.addEventListener('beforeunload', unloadHandler);
             } else if (window.attachEvent) {
                 document.attachEvent('onclick', clickHandler);
             }
@@ -116,8 +119,9 @@ var raygunRumFactory = function (window, $, Raygun) {
                 self.makePostCorsRequest(self.apiUrl + '/events?apikey=' + encodeURIComponent(self.apiKey), JSON.stringify(payload));
             }
 
-            self.sendPerformance();
+            self.sendPerformance(true, true);
             self.heartBeat();
+            self.initalStaticPageLoadTimestamp = window.performance.now();
         };
 
         this.setUser = function (user) {
@@ -139,11 +143,10 @@ var raygunRumFactory = function (window, $, Raygun) {
 
         this.heartBeat = function () {
             self.heartBeatInterval = setInterval(function () {
-
                 var data = [];
                 var payload;
 
-                extractChildData(data);
+                extractChildData(data, self.virtualPage);
 
                 if (data.length > 0) {
                     var dataJson = JSON.stringify(data);
@@ -166,11 +169,33 @@ var raygunRumFactory = function (window, $, Raygun) {
                 if (payload !== undefined) {
                     self.makePostCorsRequest(self.apiUrl + '/events?apikey=' + encodeURIComponent(self.apiKey), JSON.stringify(payload));
                 }
-            }, 30 * 1000); // 30 seconds between heartbeats
+          }, 30 * 1000); // 30 seconds between heartbeats
         };
 
-        this.sendPerformance = function () {
-            var performanceData = getPerformanceData();
+        this.virtualPageLoaded = function (path) {
+            var firstVirtualLoad = this.virtualPage == null;
+            
+            if (typeof path === 'string') {
+                if (path.length > 0 && path[0] !== '/') {
+                  path = path + '/';
+                }
+
+                this.virtualPage = path;
+            }
+            
+            if (firstVirtualLoad) {
+              this.sendPerformance(true, false);
+            } else {
+              this.sendPerformance(false, false);
+            }
+            
+            if (typeof path === 'string') {
+              this.previousVirtualPageLoadTimestamp = window.performance.now();
+            }
+        };
+
+        this.sendPerformance = function (flush, firstLoad) {
+            var performanceData = getPerformanceData(this.virtualPage, flush, firstLoad);
 
             if (performanceData === null) {
                 return;
@@ -192,8 +217,8 @@ var raygunRumFactory = function (window, $, Raygun) {
         };
 
         function stringToByteLength(str) {
-          var m = encodeURIComponent(str).match(/%[89ABab]/g);
-          return str.length + (m ? m.length : 0);
+            var m = encodeURIComponent(str).match(/%[89ABab]/g);
+            return str.length + (m ? m.length : 0);
         }
 
         function getSessionId(callback) {
@@ -311,6 +336,14 @@ var raygunRumFactory = function (window, $, Raygun) {
             return data;
         }
 
+        function generateVirtualEncodedTimingData(previousVirtualPageLoadTimestamp, initalStaticPageLoadTimestamp) {
+          return {
+            t: 'v',
+            du: Math.min(self.maxVirtualPageDuration, window.performance.now() - (previousVirtualPageLoadTimestamp || initalStaticPageLoadTimestamp)),
+            o: Math.min(self.maxVirtualPageDuration, window.performance.now() - initalStaticPageLoadTimestamp)
+          };
+        }
+
         function getEncodedTimingData(timing, offset) {
             var data = {
                 du: timing.duration,
@@ -425,15 +458,24 @@ var raygunRumFactory = function (window, $, Raygun) {
             };
         }
 
-        function getSecondaryTimingData(timing) {
+        function getVirtualPrimaryTimingData(virtualPage, previousVirtualPageLoadTimestamp, initalStaticPageLoadTimestamp) {
+            return {
+                url: window.location.protocol + '//' + window.location.host + virtualPage,
+                userAgent: navigator.userAgent,
+                timing: generateVirtualEncodedTimingData(previousVirtualPageLoadTimestamp, initalStaticPageLoadTimestamp),
+                size: 0
+            };
+        }
+
+        function getSecondaryTimingData(timing, fromZero) {
             return {
                 url: timing.name.split('?')[0],
-                timing: getSecondaryEncodedTimingData(timing, window.performance.timing.navigationStart),
+                timing: getSecondaryEncodedTimingData(timing, fromZero ? 0 : window.performance.timing.navigationStart),
                 size: timing.decodedBodySize || 0
             };
         }
 
-        function extractChildData(collection) {
+        function extractChildData(collection, fromVirtualPage) {
             if (window.performance === undefined || !window.performance.getEntries) {
                 return;
             }
@@ -469,7 +511,7 @@ var raygunRumFactory = function (window, $, Raygun) {
                         continue;
                     }
 
-                    collection.push(getSecondaryTimingData(resources[i]));
+                    collection.push(getSecondaryTimingData(resources[i], fromVirtualPage));
                 }
 
                 self.offset = resources.length;
@@ -478,16 +520,44 @@ var raygunRumFactory = function (window, $, Raygun) {
             }
         }
 
-        function getPerformanceData() {
+        function getPerformanceData(virtualPage, flush, firstLoad) {
             if (window.performance === undefined || isNaN(window.performance.timing.fetchStart)) {
                 return null;
             }
 
             var data = [];
+            
+            if (flush) {
+              // Called by the static onLoad event being fired, persist itself
+              if (firstLoad) { 
+                data.push(getPrimaryTimingData());
+              }
+              
+              // Called during both the static load event and the flush on the first virtual load call
+              extractChildData(data);  
+            }
 
-            data.push(getPrimaryTimingData());
-
-            extractChildData(data);
+            if (virtualPage) {
+              // A previous virtual load was stored, persist it and its children up until now
+              if (self.pendingVirtualPage) {
+                data.push(self.pendingVirtualPage);
+                extractChildData(data, true);
+              }
+              
+              var firstVirtualLoad = self.pendingVirtualPage == null;
+              
+              // Store the current virtual load so it can be sent upon the next one
+              self.pendingVirtualPage = getVirtualPrimaryTimingData(
+                virtualPage,
+                self.previousVirtualPageLoadTimestamp,
+                self.initalStaticPageLoadTimestamp
+              );
+              
+              // Prevent sending an empty payload for the first virtual load as we don't know when it will end
+              if (!firstVirtualLoad && data.length > 0) {
+                return data;
+              }
+            }
 
             return data;
         }
@@ -510,4 +580,4 @@ var raygunRumFactory = function (window, $, Raygun) {
     };
 };
 
-raygunRumFactory(window, window.jQuery, window.Raygun);
+raygunRumFactory(window, window.jQuery, window.__instantiatedRaygun);